[openjdk] Build with java.desktop module and X11
Revert java.desktop module build as many components in gradle depend on
awt/swing. This requires to be built with X11.

libjsound is excluded from the build to prevent dependency on ALSA, and
CUPS dependency is prevented too (but may cause issues in runtime in
Java code dependent on the sun.print package).